refactor: メソッドやモデルクラスの定義がおかしい箇所があったため修正
