+++ lib/models/record_title.dart
@@ -1,0 +1,6 @@
+class RecordOnlyTitle {
+  const RecordOnlyTitle({required this.id, required this.title, required this.createAt});
+  final int id;
+  final String title;
+  final DateTime createAt;
+}
